remove older components in favor of a single reusable recent page
