+++ packages/app/src/app/pages/Dashboard/Content/routes/Recent/index.tsx
@@ -5,57 +5,74 @@ import { Helmet } from 'react-helmet';
 import { DashboardBranch, DashboardSandbox } from 'app/pages/Dashboard/types';
 import { Loading, Stack } from '@codesandbox/components';
 import { RepoInfo } from 'app/overmind/namespaces/sidebar/types';
-import { EmptyRecent } from './EmptyRecent';
-import { RecentContent } from './RecentContent';
+import { RestrictedSandboxes } from 'app/components/StripeMessages/RestrictedSandboxes';
+import { StyledContentWrapper } from 'app/pages/Dashboard/Components/shared/elements';
+import { ContentSection } from 'app/pages/Dashboard/Components/shared/ContentSection';
+import { useWorkspaceLimits } from 'app/hooks/useWorkspaceLimits';
+import { TopBanner } from './TopBanner';
+import { CreateBranchesRow } from './CreateBranchesRow';
+import { ItemsGrid } from './ItemsGrid';
+import { EmptyCTAs } from './EmptyCTAs';
 
-export const Recent = props => {
+export const Recent = () => {
   const {
     activeTeam,
+    sidebar,
     dashboard: { sandboxes },
   } = useAppState();
+  const { isFrozen, hasReachedSandboxLimit } = useWorkspaceLimits();
   const {
     dashboard: { getPage },
   } = useActions();
+  const page = 'recent';
 
   useEffect(() => {
     getPage(sandboxesTypes.RECENT);
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [activeTeam]);
 
-  const items: (DashboardSandbox | DashboardBranch)[] =
-    sandboxes.RECENT_BRANCHES === null || sandboxes.RECENT_SANDBOXES === null
-      ? undefined
-      : [
-          ...(sandboxes.RECENT_SANDBOXES || []).map(sandbox => ({
-            type: 'sandbox' as const,
-            sandbox,
-          })),
-          ...(sandboxes.RECENT_BRANCHES || []).map(branch => ({
-            type: 'branch' as const,
-            branch,
-          })),
-        ]
-          .sort((a, b) => {
-            const dateA =
-              a.type === 'branch'
-                ? a.branch.lastAccessedAt
-                : a.sandbox.lastAccessedAt;
-            const dateB =
-              b.type === 'branch'
-                ? b.branch.lastAccessedAt
-                : b.sandbox.lastAccessedAt;
+  if (
+    sandboxes.RECENT_BRANCHES === null ||
+    sandboxes.RECENT_SANDBOXES === null
+  ) {
+    return (
+      <Stack align="center" justify="center">
+        <Loading size={12} />
+      </Stack>
+    );
+  }
 
-            return new Date(dateA) < new Date(dateB) ? 1 : -1;
-            // Merge the two data sources and show only the first 18 most recent entries
-          })
-          .slice(0, 18);
+  const recentItems: (DashboardSandbox | DashboardBranch)[] = [
+    ...(sandboxes.RECENT_SANDBOXES || []).map(sandbox => ({
+      type: 'sandbox' as const,
+      sandbox,
+    })),
+    ...(sandboxes.RECENT_BRANCHES || []).map(branch => ({
+      type: 'branch' as const,
+      branch,
+    })),
+  ]
+    .sort((a, b) => {
+      const dateA =
+        a.type === 'branch'
+          ? a.branch.lastAccessedAt
+          : a.sandbox.lastAccessedAt;
+      const dateB =
+        b.type === 'branch'
+          ? b.branch.lastAccessedAt
+          : b.sandbox.lastAccessedAt;
 
-  const branches = items?.filter(
+      return new Date(dateA) < new Date(dateB) ? 1 : -1;
+      // Merge the two data sources and show only the first 18 most recent entries
+    })
+    .slice(0, 18);
+
+  const recentBranches = recentItems.filter(
     item => item.type === 'branch'
   ) as DashboardBranch[];
 
-  const recentRepos: RepoInfo[] = branches
-    ?.map(br => ({
+  const recentRepos: RepoInfo[] = recentBranches
+    .map(br => ({
       owner: br.branch.project.repository.owner,
       name: br.branch.project.repository.name,
       defaultBranch: br.branch.project.repository.defaultBranch,
@@ -67,30 +84,54 @@ export const Recent = props => {
       return acc;
     }, []);
 
-  let pageState: 'loading' | 'ready' | 'empty';
-  if (items === undefined) {
-    pageState = 'loading';
-  } else if (items.length > 0) {
-    pageState = 'ready';
-  } else {
-    pageState = 'empty';
-  }
+  const allWorkspaceRepos = sidebar[activeTeam]?.repositories || [];
+  const otherRepos = allWorkspaceRepos.filter(
+    wr => !recentRepos.find(r => r.owner === wr.owner && r.name === wr.name)
+  );
 
   return (
-    <>
+    <StyledContentWrapper>
       <Helmet>
         <title>Recent - CodeSandbox</title>
       </Helmet>
-      {pageState === 'loading' && (
-        <Stack align="center" justify="center">
-          <Loading size={12} />
-        </Stack>
-      )}
 
-      {pageState === 'empty' && <EmptyRecent />}
-      {pageState === 'ready' && (
-        <RecentContent recentRepos={recentRepos} recentItems={items} />
+      <Stack direction="vertical" gap={4}>
+        {hasReachedSandboxLimit && <RestrictedSandboxes />}
+        <TopBanner />
+      </Stack>
+
+      <ContentSection title="Recent">
+        {recentRepos.length > 0 && (
+          <CreateBranchesRow
+            title="Create a new branch"
+            isFrozen={isFrozen}
+            repos={recentRepos}
+            trackEvent="Recent Page - Repositories - Create new branch"
+          />
+        )}
+
+        {recentItems.length > 0 ? (
+          <ItemsGrid
+            title="Pick up where you left off"
+            items={recentItems}
+            page={page}
+            activeTeam={activeTeam}
+          />
+        ) : (
+          <EmptyCTAs isFrozen={isFrozen} />
+        )}
+      </ContentSection>
+
+      {recentItems.length === 0 && otherRepos.length > 0 && (
+        <ContentSection title="Explore workspace activity">
+          <CreateBranchesRow
+            title="Start working on existing repositories"
+            repos={otherRepos}
+            isFrozen={isFrozen}
+            trackEvent="Recent Page - Explore workspace - Create new branch"
+          />
+        </ContentSection>
       )}
-    </>
+    </StyledContentWrapper>
   );
 };
